requirements.txt
pytest
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
wandb~=0.13.5
pillow~=9.3.0
dominate
scipy
pickle5